[ refact ] -> Refactoring Code & Outsourcing Routes
