+++ main.go
@@ -2,9 +2,7 @@ package main
 
 import (
 	"event-booking-api/db"
-	"event-booking-api/models"
-	"net/http"
-	"strconv"
+	"event-booking-api/routes"
 
 	"github.com/gin-gonic/gin"
 )
@@ -12,55 +10,6 @@ import (
 func main() {
 	db.InitDB()
 	server := gin.Default()
-
-	server.GET("/events", getEvents)
-	server.GET("/events/:id", getEvent)
-	server.POST("/events", createEvent)
+	routes.RegisterRoutes(server)
 	server.Run(":8080")
-}
-
-func getEvents(context *gin.Context) {
-	events, err := models.GetAllEvents()
-	if err != nil {
-		context.JSON(http.StatusInternalServerError, gin.H{"message": "could not fetch events. try again later."})
-		return
-	}
-	context.JSON(http.StatusOK, events)
-}
-
-func getEvent(context *gin.Context) {
-	eventID, err := strconv.ParseInt(context.Param("id"), 10, 64)
-	if err != nil {
-		context.JSON(http.StatusBadRequest, gin.H{"message": "could not parse event id"})
-		return
-	}
-
-	event, err := models.GetEvenByID(eventID)
-	if err != nil {
-		context.JSON(http.StatusInternalServerError, gin.H{"message": "could fetch event"})
-		return
-	}
-
-	context.JSON(http.StatusOK, event)
-}
-
-func createEvent(context *gin.Context) {
-	var event models.Event
-	err := context.ShouldBindJSON(&event)
-	if err != nil {
-		context.JSON(http.StatusBadRequest, gin.H{"message": "could not parse data"})
-		return
-	}
-
-	event.ID = 1
-	event.UserID = 1
-
-	err = event.Save()
-	if err != nil {
-		context.JSON(http.StatusInternalServerError, gin.H{"message": "could not save event. try again later"})
-		return
-	}
-
-	context.JSON(http.StatusCreated, gin.H{"message": "event created", "event": event})
-
 }
