+++ routes/events.go
@@ -1,0 +1,55 @@
+package routes
+
+import (
+	"event-booking-api/models"
+	"net/http"
+	"strconv"
+
+	"github.com/gin-gonic/gin"
+)
+
+func getEvents(context *gin.Context) {
+	events, err := models.GetAllEvents()
+	if err != nil {
+		context.JSON(http.StatusInternalServerError, gin.H{"message": "could not fetch events. try again later."})
+		return
+	}
+	context.JSON(http.StatusOK, events)
+}
+
+func getEvent(context *gin.Context) {
+	eventID, err := strconv.ParseInt(context.Param("id"), 10, 64)
+	if err != nil {
+		context.JSON(http.StatusBadRequest, gin.H{"message": "could not parse event id"})
+		return
+	}
+
+	event, err := models.GetEvenByID(eventID)
+	if err != nil {
+		context.JSON(http.StatusInternalServerError, gin.H{"message": "could fetch event"})
+		return
+	}
+
+	context.JSON(http.StatusOK, event)
+}
+
+func createEvent(context *gin.Context) {
+	var event models.Event
+	err := context.ShouldBindJSON(&event)
+	if err != nil {
+		context.JSON(http.StatusBadRequest, gin.H{"message": "could not parse data"})
+		return
+	}
+
+	event.ID = 1
+	event.UserID = 1
+
+	err = event.Save()
+	if err != nil {
+		context.JSON(http.StatusInternalServerError, gin.H{"message": "could not save event. try again later"})
+		return
+	}
+
+	context.JSON(http.StatusCreated, gin.H{"message": "event created", "event": event})
+
+}
