+++ routes/routes.go
@@ -1,0 +1,9 @@
+package routes
+
+import "github.com/gin-gonic/gin"
+
+func RegisterRoutes(server *gin.Engine) {
+	server.GET("/events", getEvents)
+	server.GET("/events/:id", getEvent)
+	server.POST("/events", createEvent)
+}
